Make `MaybeValueReference::value` less confusing
The function that gets the associated value of a `MaybeValueReference`
is called "value", and so is the function used to create a
`MaybeValueReference` containing a value. In order to fix this, the
function getting the value has been renamed to "as_value". This also
makes it more consistent with `Value` itself.